refractor code
- dont save personal weapon list
- restore persist_weapon
- fix pv list does not save for second character
- add desync protection infraction
- network player manager should reset known players and desyncs
- fix language value not signed (-1 not saved) for bad players

+++ src/services/bad_players/bad_players.cpp
@@ -25,7 +25,7 @@ namespace big
 	{
 		auto& bs = scr_globals::gpbd_fm_3.as<GPBD_FM_3*>()->Entries[player->id()].BossGoon;
 		add_player(player->m_rockstar_id,
-		    {player->m_name, block_join, is_spammer, (u_int)bs.Language, is_spammer ? "spam" : player->spam_message});
+		    {player->m_name, block_join, is_spammer, bs.Language, is_spammer ? "spam" : player->spam_message});
 	}
 
 	void bad_players_service::toggle_block(rock_id rockstar_id, bool v)
